refactor: split chat.cpp to reduce compile time
